requirements.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.6
cycler==0.11.0
Deprecated==1.2.13
Django==3.2.13
django-celery-beat==2.2.1
django-celery-results==2.3.1
django-cors-headers==3.12.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
fonttools==4.38.0
idna==3.4
importlib-metadata==4.11.4
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.6.0
multitasking==0.0.11
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
requests==2.28.1
seaborn==0.12.1
six==1.16.0
sqlparse==0.4.2
twilio==7.9.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yfinance==0.1.84
zipp==3.8.0
